--- tests/test_parser.py
@@ -1,68 +1,0 @@
-from parsero.common.errors import LexicalError, SyntacticError
-from parsero.parser import Parser
-
-# If a language fails to parse it will raise an error
-
-
-def test_python():
-    parser = Parser(f"examples/python/python.regex", f"examples/python/python.ghm")
-    parser.parse(f"examples/python/example.py")
-
-    try:
-        parser.parse(f"examples/c/example.c")
-    except LexicalError:
-        pass
-    except SyntacticError:
-        pass
-    else:
-        # If it can compile wrong language raise error
-        assert False
-
-
-def test_c():
-    parser = Parser(f"examples/c/c.regex", f"examples/c/c.ghm")
-    parser.parse(f"examples/c/example.c")
-
-    try:
-        parser.parse(f"examples/lisp/example.lisp")
-    except LexicalError:
-        pass
-    except SyntacticError:
-        pass
-    else:
-        # If it can compile wrong language raise error
-        assert False
-
-
-def test_lisp():
-    parser = Parser(f"examples/lisp/lisp.regex", f"examples/lisp/lisp.ghm")
-    parser.parse(f"examples/lisp/example.lisp")
-
-    try:
-        parser.parse(f"examples/json/example.json")
-    except LexicalError:
-        pass
-    except SyntacticError:
-        pass
-    else:
-        # If it can compile wrong language raise error
-        assert False
-
-
-def test_json():
-    parser = Parser(f"examples/json/json.regex", f"examples/json/json.ghm")
-    parser.parse(f"examples/json/example.json")
-
-    try:
-        parser.parse(f"examples/python/example.py")
-    except LexicalError:
-        pass
-    except SyntacticError:
-        pass
-    else:
-        # If it can compile wrong language raise error
-        assert False
-
-
-if __name__ == "__main__":
-    test_python()
